# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ChatMod Mobile API
  version: 0.1.0
  description: Backend contract for the phone-hosted YouTube Live moderation bot.
servers:
  - url: http://localhost:4100
paths:
...
  /timers:
    get:
      summary: List timed messages
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Timer list
    post:
      summary: Create timed message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimerInput"
      responses:
        "201":
          description: Timer created
        "403":
          description: Current plan timed-message limit reached
  /timers/due:
    post:
      summary: Find timed messages due to send
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DueTimerRequest"
      responses:
        "200":
          description: Due timer list
  /timers/{id}:
    put:
      summary: Create or replace a timed message using a stable client id
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimerInput"
      responses:
        "200":
          description: Timer upserted
        "403":
          description: Current plan timed-message limit reached
    patch:
      summary: Update timed message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TimerPatch"
      responses:
        "200":
          description: Timer updated
    delete:
      summary: Delete timed message
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Timer deleted
  /timers/{id}/mark-sent:
    post:
      summary: Mark a timed message as sent
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Timer updated
...
  /team/profiles/{profileId}/members:
    get:
      summary: List invited, active, and revoked team moderators for a channel profile
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile team member list and seat limits
          content:
            application/json:
              schema:
                type: object
                required:
                  - profileId
                  - teamSeats
                  - extraSeats
                  - members
                properties:
                  profileId:
                    type: string
                  teamSeats:
                    type: integer
                  extraSeats:
                    type: integer
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeamMember"
  /team/profiles/{profileId}/invites:
    post:
      summary: Create a plan-gated team moderator invite for a channel profile
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInviteCreate"
      responses:
        "201":
          description: Team invite created; invite code is returned once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamInviteResult"
        "403":
          description: Team moderator seat limit reached for current plan
  /team/profiles/{profileId}/members/{memberId}:
    delete:
      summary: Revoke a team moderator invite or active membership
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
        - in: path
          name: memberId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Team member revoked
          content:
            application/json:
              schema:
                type: object
                required:
                  - member
                properties:
                  member:
                    $ref: "#/components/schemas/TeamMember"
  /team/invites/redeem:
    post:
      summary: Redeem a team moderator invite on another Android device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInviteRedeem"
      responses:
        "200":
          description: Team membership activated for the current device
          content:
            application/json:
              schema:
                type: object
                required:
                  - membership
                properties:
                  membership:
                    $ref: "#/components/schemas/TeamMembership"
        "409":
          description: Invite has already been redeemed by another device
  /team/memberships:
    get:
      summary: List team memberships redeemed by the current device
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active team memberships for this device
          content:
            application/json:
              schema:
                type: object
                required:
                  - memberships
                properties:
                  memberships:
                    type: array
                    items:
                      $ref: "#/components/schemas/TeamMembership"
...
  /integrations/discord/test:
    post:
      summary: Send a test Discord alert for a configured profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - profileId
              properties:
                profileId:
                  type: string
      responses:
        "200":
          description: Test alert result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscordAlertResult"
        "403":
          description: Discord alerts require Pro or Creator plan
        "409":
          description: Discord webhook is not configured or enabled
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminApiKey:
      type: apiKey
      in: header
      name: x-admin-api-key
  schemas:
...
    TeamMemberPermissions:
      type: object
      required:
        - viewQueue
        - moderate
        - manageWarnings
        - viewAnalytics
      properties:
        viewQueue:
          type: boolean
        moderate:
          type: boolean
        manageWarnings:
          type: boolean
        viewAnalytics:
          type: boolean
    TeamInviteCreate:
      type: object
      required:
        - displayName
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        role:
          type: string
          enum:
            - moderator
            - producer
          default: moderator
        permissions:
          $ref: "#/components/schemas/TeamMemberPermissions"
    TeamInviteRedeem:
      type: object
      required:
        - inviteCode
      properties:
        inviteCode:
          type: string
          minLength: 12
          maxLength: 120
        displayName:
          type: string
          minLength: 1
          maxLength: 80
    TeamMember:
      type: object
      required:
        - id
        - profileId
        - displayName
        - role
        - status
        - inviteCodePreview
        - memberDeviceId
        - permissions
        - acceptedAt
        - revokedAt
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        profileId:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum:
            - moderator
            - producer
        status:
          type: string
          enum:
            - invited
            - active
            - revoked
        inviteCodePreview:
          type: string
        memberDeviceId:
          type:
            - string
            - "null"
        permissions:
          $ref: "#/components/schemas/TeamMemberPermissions"
        acceptedAt:
          type:
            - string
            - "null"
          format: date-time
        revokedAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeamInviteResult:
      type: object
      required:
        - member
        - inviteCode
      properties:
        member:
          $ref: "#/components/schemas/TeamMember"
        inviteCode:
          description: Full invite code returned only at creation time. Store or share it immediately.
          type: string
    TeamMembership:
      allOf:
        - $ref: "#/components/schemas/TeamMember"
        - type: object
          required:
            - profileName
            - channelId
            - ownerDeviceId
          properties:
            profileName:
              type: string
            channelId:
              type: string
            ownerDeviceId:
              type: string
...
    TimerInput:
      type: object
      required:
        - profileId
        - name
        - message
        - intervalMinutes
      properties:
        profileId:
          type: string
        name:
          type: string
          maxLength: 80
        message:
          type: string
          maxLength: 500
        intervalMinutes:
          type: integer
          minimum: 1
          maximum: 240
        minChatMessages:
          type: integer
          minimum: 0
          maximum: 500
        quietStartMinutes:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 1440
        quietEndMinutes:
          type:
            - integer
            - "null"
          minimum: 1
          maximum: 1440
        enabled:
          type: boolean
    TimerPatch:
      type: object
      properties:
        name:
          type: string
          maxLength: 80
        message:
          type: string
          maxLength: 500
        intervalMinutes:
          type: integer
          minimum: 1
          maximum: 240
        minChatMessages:
          type: integer
          minimum: 0
          maximum: 500
        quietStartMinutes:
          type:
            - integer
            - "null"
          minimum: 0
          maximum: 1440
        quietEndMinutes:
          type:
            - integer
            - "null"
          minimum: 1
          maximum: 1440
        enabled:
          type: boolean
    DueTimerRequest:
      type: object
      required:
        - profileId
        - messagesSinceLastTimer
      properties:
        profileId:
          type: string
        messagesSinceLastTimer:
          type: integer
          minimum: 0
        now:
          type: string
          format: date-time
        streamStartedAt:
          type: string
          format: date-time
...
    DiscordAlertResult:
      type: object
      required:
        - sent
        - profileId
      properties:
        sent:
          type: boolean
        skippedReason:
          type:
            - string
            - "null"
          enum:
            - not_configured
            - disabled
            - moderation_alerts_disabled
            - runtime_alerts_disabled
            - null
        sentAt:
          type:
            - string
            - "null"
          format: date-time
        profileId:
          type: string
...